Move options from consumer to ingester
